+++ apps/studio/components/interfaces/Markdown.tsx
@@ -1,8 +1,7 @@
-import { ReactMarkdownProps } from 'react-markdown/lib/complex-types'
-import { ReactMarkdown } from 'react-markdown/lib/react-markdown'
+import { ReactMarkdown, ReactMarkdownOptions } from 'react-markdown/lib/react-markdown'
 import { cn } from 'ui'
 
-interface Props extends Omit<ReactMarkdownProps, 'children' | 'node'> {
+interface Props extends Omit<ReactMarkdownOptions, 'children' | 'node'> {
   className?: string
   content: string
   extLinks?: boolean
