+++ apps/studio/pages/_app.tsx
@@ -14,6 +14,7 @@ import 'styles/stripe.scss'
 import 'styles/toast.scss'
 import 'styles/ui.scss'
 import 'styles/reactflow.scss'
+import 'styles/markdown-preview.scss'
 
 import { loader } from '@monaco-editor/react'
 import { TooltipProvider } from '@radix-ui/react-tooltip'
